+++ src/error/include/error/base_error.hpp
@@ -73,6 +73,7 @@ class Error
     std::vector<Error> _subErrors;
 
 #ifndef NDEBUG
+    /// The source location where the error was created, used for debugging
     std::source_location _location;
 #endif
 
@@ -101,9 +102,6 @@ class Error
         const std::optional<std::string>& newMessage  = std::nullopt,
         bool                              addSubError = true
     ) const;
-
-    [[nodiscard]]
-    static Error NotYetImplemented();
 
    protected:
     [[nodiscard]] const std::string& getMessage() const;